This program extends the previous one by adding functionality for performing join operations between tables. It also includes some improvements and bug fixes
Join Tables:

Added a join_tables function to perform join operations between two tables based on a specified column.
Modified the select function to handle join operations. Now, when a user wants to select data from a table, they can also specify another table to join with and the column to join on.
Regular Expressions:

Used regular expressions to parse the CREATE TABLE statement, extracting the table name and column names.
This change makes the parsing of CREATE TABLE statements more robust and allows for better handling of different column definitions.
Bug Fixes:

Fixed a bug in the delete function where the row index for the column to be deleted was not correctly calculated.
Fixed a bug in the update function where the updated row was not correctly written back to the JSON file.